Refactor tools and clean setup
