+++ AGENTS/tools/dump_headers.py
