+++ tests/conftest.py
@@ -174,7 +174,7 @@ Active Faculty for this Session:
     
     root_logger.info("--------------------------------------------------------------------------------")
     # --- End Enhanced Log Header & Faculty Information ---
-    from dump_headers import dump_headers
+    from AGENTS.tools.dump_headers import dump_headers
     buf = StringIO()
     _stdout = sys.stdout
     sys.stdout = buf
